chore(crons): Rename badly named api file
I accidentally pluralised this when creating it, just fixing.
