+++ src/sentry/monitors/endpoints/project_monitor_details.py
